+++ pyxel/data_structure/pixel.py
@@ -26,7 +26,7 @@ class Pixel(Array):
     Accepted array types: np.int32, np.int64, np.uint32, np.uint64, np.float16, np.float32, np.float64.
     """
 
-    EXP_TYPE = np.int
+    EXP_TYPE = int
     TYPE_LIST = (
         np.int32,
         np.int64,
